# src/main/resources/openapi/openapi.yml
openapi: 3.0.3
info:
  title: Vet API
  version: 1.0.0
  description: API to manage veterinarians in the Pet Clinic system.
  termsOfService: https://example.com/terms
  contact:
    name: API Support
    url: https://example.com/support
    email: support@example.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  #$ref: './api-info.yml'
paths:
  #same endpoints with different HTTP methods, must be placed under same path. (example /vet GET/POST)
  /vet:
    get:
      summary: Retrieves a paginated list of veterinarians
      description: |
        Returns a paginated list of all registered veterinarians. Supports
        `page` and `size` query parameters for pagination.
      tags:
        - Vet
      operationId: vetDetails
      parameters:
        - name: page
          in: query
          description: Zero-based page index
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: size
          in: query
          description: Maximum number of items per page
          schema:
            type: integer
            default: 10
            minimum: 1
      responses:
        '200':
          description: List of veterinarians returned successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VetResponse'
    post:
      summary: Create a new veterinarian
      tags:
        - Vet
      operationId: createVet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateVetRequest'
      responses:
        '200':
          description: Created vet
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VetResponse'

  /vet/{lastName}/{specialtyNames}:
    get:
      summary: Find vets by last name and specialties
      tags:
        - Vet
      operationId: findByLastNameAndSpecialties
      parameters:
        - name: lastName
          in: path
          required: true
          schema:
            type: string
        - name: specialtyNames
          in: path
          required: true
          schema:
            type: array
            items:
              type: string
            style: simple
      responses:
        '200':
          description: List of matching vets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: './models/components-schemas.yml#/components/schemas/VetResponse'

  /vet/{firstName}/{lastName}: #same endpoints with different HTTP methods, must be placed under same path.
    delete:
      summary: Delete vet by name
      tags:
        - Vet
      operationId: deleteByName
      parameters:
        - name: firstName
          in: path
          required: true
          schema:
            type: string
        - name: lastName
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Vet deleted

    put:
      summary: Update vet details
      tags:
        - Vet
      operationId: updateVet
      parameters:
        - name: firstName
          in: path
          required: true
          schema:
            type: string
        - name: lastName
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 10
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './models/components-schemas.yml#/components/schemas/UpdateVetRequest'
      responses:
        '200':
          description: Updated vet
          content:
            application/json:
              schema:
                $ref: './models/components-schemas.yml#/components/schemas/VetResponse'
#Need to be defined here in order to be present in Swagger-UI
components:
  schemas:
    CreateVetRequest:
      $ref: './models/components-schemas.yml#/components/schemas/CreateVetRequest'
    UpdateVetRequest:
      $ref: './models/components-schemas.yml#/components/schemas/UpdateVetRequest'
    VetResponse:
      $ref: './models/components-schemas.yml#/components/schemas/VetResponse'